pyproject.toml
# 项目配置文件
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "web-automation-framework"
version = "1.0.0"
description = "Web Automation Test Framework using Playwright with Python"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["automation", "testing", "playwright", "pytest", "web"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "playwright>=1.48.0",
    "pytest>=8.3.4",
    "pytest-html>=4.1.1",
    "pytest-xdist>=3.6.1",
    "pytest-rerunfailures>=14.0",
    "pytest-base-url>=2.1.0",
    "allure-pytest>=2.13.5",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-timeout>=2.3.1",
    "pytest-randomly>=3.16.0",
    "faker>=33.1.0",
    "black>=24.10.0",
    "flake8>=7.1.1",
    "pylint>=3.3.2",
    "mypy>=1.14.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/web-automation-framework"
Repository = "https://github.com/yourusername/web-automation-framework"
Issues = "https://github.com/yourusername/web-automation-framework/issues"

[project.scripts]
web-automation = "run_tests:main"

# Black 配置
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | venv
)/
'''

# Pytest 配置
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*", "*Tests"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--html=reports/html/report.html",
    "--self-contained-html",
    "--alluredir=reports/allure",
    "--junitxml=reports/junit/results.xml",
]
markers = [
    "smoke: Quick smoke tests",
    "regression: Full regression tests",
    "sanity: Sanity tests",
    "api: API tests",
    "ui: UI tests",
    "slow: Slow running tests",
    "flaky: Tests that may occasionally fail",
    "skip: Skip these tests",
    "critical: Critical business flows",
]

# Coverage 配置
[tool.coverage.run]
source = ["pages", "utils", "elements"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

# MyPy 配置
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

# Pylint 配置
[tool.pylint.messages_control]
disable = [
    "C0111", # missing-docstring
    "C0103", # invalid-name
    "R0903", # too-few-public-methods
]

[tool.pylint.format]
max-line-length = 100
